Exercise 4.21- change the delete blog operation so that a blog can only be deleted by the user who added it

+++ part4/bloglist/controllers/blogs.js
@@ -57,8 +57,27 @@ blogsRouter.get('/:id', async (request, response) => {
 
 // Identify each resource with a unique URL following RESTful convention
 blogsRouter.delete('/:id', async (request, response) => {
-    await Blog.findByIdAndDelete(request.params.id)
-    response.status(204).end()
+    const decodedToken = jwt.verify(request.token, process.env.SECRET)
+    if (!decodedToken.id) {
+        return response.status(401).json({ error: 'token invalid' })
+    }
+
+    const userId = decodedToken.id
+    const blog = await Blog.findById(request.params.id)
+
+    // Check if the blog exists
+    if (!blog) {
+        return response.status(404).json({ error: 'blog not found' });
+    }
+
+    // Verify if the blog was created by the requesting user
+    if (blog.user.toString() !== userId) {
+        return response.status(403).json({ error: 'permission denied' });
+    }
+
+    // Delete the blog if all checks pass
+    await Blog.findByIdAndDelete(request.params.id);
+    response.status(204).end();
 })
 
 // Update the information for a specific blog post
